pyproject.toml
[tool.poetry]
name = "zycelium-zygote"
version = "0.0.1"
description = "Personal Zycelium Instance"
authors = ["Harshad Sharma <harshad@sharma.io>"]
license = "BSD 3-Clause License"
readme = "README.md"
packages = [{include = "zycelium", from = "src"}]

[tool.poetry.scripts]
zygote = "zycelium.zygote.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
quart = "^0.18.4"
quart-cors = "^0.6.0"
async-signals = "^0.1.5"
tortoise-orm = "^0.19.3"
pydantic = "^1.10.7"
quart-auth = "^0.8.0"
python-socketio = "^5.8.0"
apscheduler = "^3.10.1"
zycelium-dataconfig = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
